+++ sketch_oxide/src/graph/mod.rs
@@ -6,6 +6,7 @@
 mod agm_connectivity;
 mod doulion;
 mod gss;
+mod hyperanf;
 mod mascot;
 mod tcm;
 mod triest;
@@ -13,6 +14,7 @@ mod triest;
 pub use agm_connectivity::AgmConnectivity;
 pub use doulion::Doulion;
 pub use gss::GssSketch;
+pub use hyperanf::HyperAnf;
 pub use mascot::Mascot;
 pub use tcm::TcmSketch;
 pub use triest::Triest;